LayoutTests/media/modern-media-controls/media-controls/media-controls-placard.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-item.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-service.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/start-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/play-pause-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/skip-back-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/airplay-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/pip-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/fullscreen-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/slider.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scrubber.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/time-label.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/time-control.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/placard.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/airplay-placard.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/pip-placard.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/media-controls.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>MediaControls</code> handling of placards.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

const mediaControls = new MediaControls({ layoutTraits: LayoutTraits.macOS });

debug("By default, no placard is shown.");
shouldBeFalse("mediaControls.showsPlacard");
shouldBe("mediaControls.children", "[mediaControls.controlsBar]");

debug("");
debug("Showing the Airplay placard");
mediaControls.showPlacard(mediaControls.airplayPlacard);
shouldBeTrue("mediaControls.showsPlacard");
shouldBe("mediaControls.children", "[mediaControls.airplayPlacard, mediaControls.controlsBar]");

debug("");
debug("Showing the PiP placard, hides the controls bar");
mediaControls.showPlacard(mediaControls.pipPlacard);
shouldBeTrue("mediaControls.showsPlacard");
shouldBe("mediaControls.children", "[mediaControls.pipPlacard]");

debug("");
debug("Hide placard");
mediaControls.hidePlacard();
shouldBeFalse("mediaControls.showsPlacard");
shouldBe("mediaControls.children", "[mediaControls.controlsBar]");

debug("");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
